refactored batching into custom dao implementation.
