+++ api/src/main/java/org/pmiops/workbench/cohortreview/CohortReviewServiceImpl.java
@@ -12,32 +12,19 @@ import org.pmiops.workbench.exceptions.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Slice;
-import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.sql.Connection;
-import java.sql.SQLException;
-import java.sql.Statement;
 import java.util.List;
-import java.util.logging.Level;
 import java.util.logging.Logger;
 
 @Service
 public class CohortReviewServiceImpl implements CohortReviewService {
 
-    private static final String SQL_TEMPLATE = "insert into participant_cohort_status(" +
-            "birth_date, ethnicity_concept_id, gender_concept_id, race_concept_id, " +
-            "status, cohort_review_id, participant_id) " +
-            "values";
-    private static final String NEXT_INSERT = " (%s, %d, %d, %d, %d, %d, %d)";
-    private static final int BATCH_SIZE = 50;
-
     private CohortReviewDao cohortReviewDao;
     private CohortDao cohortDao;
     private ParticipantCohortStatusDao participantCohortStatusDao;
     private WorkspaceService workspaceService;
-    private JdbcTemplate jdbcTemplate;
 
     private static final Logger log = Logger.getLogger(CohortReviewServiceImpl.class.getName());
 
@@ -45,13 +32,11 @@ public class CohortReviewServiceImpl implements CohortReviewService {
     CohortReviewServiceImpl(CohortReviewDao cohortReviewDao,
                             CohortDao cohortDao,
                             ParticipantCohortStatusDao participantCohortStatusDao,
-                            WorkspaceService workspaceService,
-                            JdbcTemplate jdbcTemplate) {
+                            WorkspaceService workspaceService) {
         this.cohortReviewDao = cohortReviewDao;
         this.cohortDao = cohortDao;
         this.participantCohortStatusDao = participantCohortStatusDao;
         this.workspaceService = workspaceService;
-        this.jdbcTemplate = jdbcTemplate;
     }
 
     public CohortReviewServiceImpl() {
@@ -110,61 +95,7 @@ public class CohortReviewServiceImpl implements CohortReviewService {
     @Transactional
     public void saveFullCohortReview(CohortReview cohortReview, List<ParticipantCohortStatus> participantCohortStatuses) {
         cohortReview = saveCohortReview(cohortReview);
-        saveParticipantCohortStatuses(participantCohortStatuses);
-    }
-
-    /**
-     * This implementation manually creates batched sql statements. For unknown reasons Spring JPA nor
-     * JDBC batching works in appengine running in the cloud. It's old school but it solves our batching
-     * issue and can handle inserts up to 10,000 participants.
-     */
-    protected void saveParticipantCohortStatuses(List<ParticipantCohortStatus> participantCohortStatuses) {
-        Statement statement = null;
-        Connection connection = null;
-        int index = 0;
-        String sqlStatement = SQL_TEMPLATE;
-
-        try {
-            connection = jdbcTemplate.getDataSource().getConnection();
-            statement = connection.createStatement();
-            connection.setAutoCommit(false);
-
-            for (ParticipantCohortStatus pcs : participantCohortStatuses) {
-                String birthDate = pcs.getBirthDate() == null
-                        ? "NULL" : "'" + pcs.getBirthDate().toString() + "'";
-                String nextSql = String.format(NEXT_INSERT,
-                        birthDate,
-                        pcs.getEthnicityConceptId(),
-                        pcs.getGenderConceptId(),
-                        pcs.getRaceConceptId(),
-                        //this represents NOT_REVIEWED
-                        3,
-                        pcs.getParticipantKey().getCohortReviewId(),
-                        pcs.getParticipantKey().getParticipantId());
-                sqlStatement = sqlStatement.equals(SQL_TEMPLATE)
-                        ? sqlStatement + nextSql : sqlStatement + ", " + nextSql;
-
-                if(++index % BATCH_SIZE == 0) {
-                    statement.execute(sqlStatement);
-                    sqlStatement = SQL_TEMPLATE;
-                }
-            }
-
-            if (!sqlStatement.equals(SQL_TEMPLATE)) {
-                statement.execute(sqlStatement);
-            }
-
-            connection.commit();
-            connection.setAutoCommit(true);
-
-        } catch (SQLException ex) {
-            log.log(Level.INFO, "SQLException: " + ex.getMessage());
-            rollback(connection);
-            throw new RuntimeException("SQLException: " + ex.getMessage(), ex);
-        } finally {
-            close(statement);
-            close(connection);
-        }
+        participantCohortStatusDao.saveParticipantCohortStatusesCustom(participantCohortStatuses);
     }
 
     @Override
@@ -189,44 +120,5 @@ public class CohortReviewServiceImpl implements CohortReviewService {
     @Override
     public Slice<ParticipantCohortStatus> findParticipantCohortStatuses(Long cohortReviewId, PageRequest pageRequest) {
         return participantCohortStatusDao.findByParticipantKey_CohortReviewId(cohortReviewId, pageRequest);
-    }
-
-    /**
-     * This doesn't actually close the pooled connection, but is more likely to return this connection
-     * back to the connection pool for reuse.
-     *
-     * @param connection
-     */
-    private void close(Connection connection) {
-        if (connection != null) {
-            try {
-                connection.close();
-            } catch (SQLException e) {
-                log.log(Level.INFO, "Problem closing connection: " + e.getMessage());
-                throw new RuntimeException("SQLException: " + e.getMessage(), e);
-            }
-        }
-    }
-
-    private void close(Statement statement) {
-        if (statement != null) {
-            try {
-                statement.close();
-            } catch (SQLException e) {
-                log.log(Level.INFO, "Problem closing prepared statement: " + e.getMessage());
-                throw new RuntimeException("SQLException: " + e.getMessage(), e);
-            }
-        }
-    }
-
-    private void rollback(Connection connection) {
-        if (connection != null) {
-            try {
-                connection.rollback();
-            } catch (SQLException e) {
-                log.log(Level.INFO, "Problem on rollback: " + e.getMessage());
-                throw new RuntimeException("SQLException: " + e.getMessage(), e);
-            }
-        }
     }
 }
