+++ api/src/test/java/org/pmiops/workbench/db/dao/ParticipantCohortStatusDaoTest.java
@@ -21,12 +21,13 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.sql.Date;
 import java.util.Arrays;
+import java.util.List;
 
 import static org.junit.Assert.assertEquals;
 
 @RunWith(SpringRunner.class)
 @DataJpaTest
-@Import(LiquibaseAutoConfiguration.class)
+@Import({LiquibaseAutoConfiguration.class})
 @AutoConfigureTestDatabase(replace= AutoConfigureTestDatabase.Replace.NONE)
 @Transactional
 public class ParticipantCohortStatusDaoTest {
@@ -45,14 +46,14 @@ public class ParticipantCohortStatusDaoTest {
         participant1 = new ParticipantCohortStatus()
                 .participantKey(key1)
                 .status(CohortStatus.INCLUDED)
-                .birthDate(new Date(-852058800000000L))
+                .birthDate(new Date(System.currentTimeMillis()))
                 .ethnicityConceptId(1L)
                 .genderConceptId(1L)
                 .raceConceptId(1L);
         participant2 = new ParticipantCohortStatus()
                 .participantKey(key2)
                 .status(CohortStatus.EXCLUDED)
-                .birthDate(new Date(-852058800000000L))
+                .birthDate(new Date(System.currentTimeMillis()))
                 .ethnicityConceptId(1L)
                 .genderConceptId(1L)
                 .raceConceptId(1L);
@@ -93,6 +94,31 @@ public class ParticipantCohortStatusDaoTest {
                         COHORT_REVIEW_ID,
                         participant1.getParticipantKey().getParticipantId()
                 ));
+    }
+
+    @Test
+    public void saveParticipantCohortStatuses() throws Exception {
+        ParticipantCohortStatusKey key1 = new ParticipantCohortStatusKey().cohortReviewId(2).participantId(3);
+        ParticipantCohortStatusKey key2 = new ParticipantCohortStatusKey().cohortReviewId(2).participantId(4);
+        ParticipantCohortStatus pcs1 = new ParticipantCohortStatus()
+                .participantKey(key1)
+                .status(CohortStatus.INCLUDED)
+                .birthDate(new Date(System.currentTimeMillis()))
+                .ethnicityConceptId(1L)
+                .genderConceptId(1L)
+                .raceConceptId(1L);
+        ParticipantCohortStatus pcs2 = new ParticipantCohortStatus()
+                .participantKey(key2)
+                .status(CohortStatus.EXCLUDED)
+                .birthDate(new Date(System.currentTimeMillis()))
+                .ethnicityConceptId(1L)
+                .genderConceptId(1L)
+                .raceConceptId(1L);
+
+        participantCohortStatusDao.saveParticipantCohortStatusesCustom(Arrays.asList(pcs1, pcs2));
+        List<ParticipantCohortStatus> pcdList =
+                participantCohortStatusDao.findByParticipantKey_CohortReviewId(2, new PageRequest(0, 10)).getContent();
+        assertEquals(2, pcdList.size());
     }
 
     private void assertParticipant(Pageable pageRequest, ParticipantCohortStatus expectedParticipant) {
